スピードバーの表示

+++ app/src/main/java/jp/ceed/android/mylapslogger/network/response/SessionsResponse.java
@@ -263,7 +263,7 @@ public class SessionsResponse implements Serializable {
 				 */
 
 				@SerializedName("kph")
-				public double kph;
+				public float kph;
 
 				@SerializedName("mps")
 				public double mps;
